+++ src/main/kotlin/lotto/Controller/PrizeUtils.kt
@@ -1,0 +1,9 @@
+package lotto.Controller
+
+import lotto.Model.Prize
+
+fun getPrizesToPrint(): List<Prize> {
+    return Prize.values()
+        .filter { it.matchCount >= 3 }
+        .sortedBy { it.matchCount }
+}
